Added translation and rotation functions

+++ del.py
@@ -2,5 +2,7 @@ from pathlib import Path
 
 for ext in ("*.ppm", "*.png"):
     for file in Path(".").rglob(ext):
+        if "assets" in file.parts:
+            continue
         file.unlink()
         print(f"Deleted {file}")
